video-driver: check session support after setting resolution
For encoder case, the codec is not started until the first frame is queued. So
current session should be checked for support after setting resolution. Otherwise,
when the resolution is too large to support the current session. For frameworks,
resource reclamation is currently not triggered due to the queue buffer process.

Change-Id: Ia38deb0543dbeef6436c805e4e8c20b463f143d5
Signed-off-by: Ruchika Ashtankar <quic_rashtank@quicinc.com>