video-driver: Fix AV1 SEF + FG buffer issues
On old chipsets for AV1+SEF+FG, no-show frame need to be decoded, and
then SEF frame (show_existing_frame=1) needs to be decoded as well.
Because no-show frame's OPB is returned to host immediately after
decoding, only DPB is kept which doesn't have film-grain information.
The SEF frame is just a placeholder to display that no-show frame, so
it's double decoding for one shown frame. The new change in FW is to
hold the OPB of no-show frame until the SEF frame comes, or if there's
no SEF frame coming, wait until it finally got released from DPB list.
Then SEF frame doesn't need to be decoded. Because there could be as
many no-show frames until a SEF frame comes, so the OPBs held by FW will
accumulate. the worst case is all 11 OPBs are kept for no-show frames in
FW, then one of them will be released from DPB list.

Change-Id: If245c7412a433197621b7659384cb61174efd291
Signed-off-by: Manikanta Kanamarlapudi <quic_kmanikan@quicinc.com>