video: driver: updated 1080p session resolution check
In current code base the resolution check for standard
1080p fails and will not be consider as 1080p session,
due to which we are able to run more number of 1080p
sessions than the expected one.

The current logic of 1.5 times resolution check will be
ideal for resolutions higher that 720p where memory will
be a constraint, for standard HD and full HD resolutions
we can go ahead with the default resolution check instead
of 1.5 times resolution.

Change-Id: I3ba9ac879c2c2ee4068ce95f942e1ffaad1631be
Signed-off-by: RAJA SHEKAR J <quic_rshekarj@quicinc.com>
Signed-off-by: Shivam Joshi <quic_shivamj@quicinc.com>